reuse existing tests wherever possible
